+++ src/webgl/shaderity_shaders/SummedAreaTableShader/SummedAreaTableShader.vert
@@ -17,8 +17,8 @@ void main()
 /* shaderity: @{mainPrerequisites} */
 
   mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
-  mat4 viewMatrix = get_viewMatrix(uint(cameraSID));
-  mat4 projectionMatrix = get_projectionMatrix(uint(cameraSID));
+  mat4 viewMatrix = get_viewMatrix(cameraSID);
+  mat4 projectionMatrix = get_projectionMatrix(cameraSID);
   gl_Position = projectionMatrix * viewMatrix * worldMatrix * a_position;
 
 v_texcoord_0 = a_texcoord_0;
