+++ src/webgl/shaderity_shaders/common/iblDefinition.glsl
@@ -69,7 +69,7 @@ vec3 get_sample_from_backbuffer(vec2 sampleCoord, float perceptualRoughness, flo
 
 vec3 getIBLVolumeRefraction(vec3 baseColor, vec3 normal, vec3 view, float cameraSID, float materialSID, float thickness, float perceptualRoughness, float ior, vec3 attenuationColor, float attenuationDistance) {
 #ifdef RN_USE_DISPERSION
-  float dispersion = get_dispersion(uint(materialSID), 0u);
+  float dispersion = get_dispersion(materialSID, 0u);
   float halfSpread = (ior - 1.0) * 0.025 * dispersion;
   vec3 iors = vec3(ior - halfSpread, ior, ior + halfSpread);
 
@@ -80,7 +80,7 @@ vec3 getIBLVolumeRefraction(vec3 baseColor, vec3 normal, vec3 view, float camera
     transmissionRayLength = length(transmissionRay);
     vec3 refractedRayExit = v_position_inWorld.xyz + transmissionRay;
 
-    vec4 ndcPos = get_projectionMatrix(uint(cameraSID)) * get_viewMatrix(uint(cameraSID)) * vec4(refractedRayExit, 1.0);
+    vec4 ndcPos = get_projectionMatrix(cameraSID) * get_viewMatrix(cameraSID) * vec4(refractedRayExit, 1.0);
     vec2 refractionCoords = ndcPos.xy / ndcPos.w;
     refractionCoords += 1.0;
     refractionCoords /= 2.0;
@@ -92,7 +92,7 @@ vec3 getIBLVolumeRefraction(vec3 baseColor, vec3 normal, vec3 view, float camera
   float transmissionRayLength = length(transmissionRay);
   vec3 refractedRayExit = v_position_inWorld.xyz + transmissionRay;
 
-  vec4 ndcPos = get_projectionMatrix(uint(cameraSID)) * get_viewMatrix(uint(cameraSID)) * vec4(refractedRayExit, 1.0);
+  vec4 ndcPos = get_projectionMatrix(cameraSID) * get_viewMatrix(cameraSID) * vec4(refractedRayExit, 1.0);
   vec2 refractionCoords = ndcPos.xy / ndcPos.w;
   refractionCoords += 1.0;
   refractionCoords /= 2.0;
@@ -199,7 +199,7 @@ vec3 sheenIBL(float NdotV, float sheenPerceptualRoughness, vec3 sheenColor, vec4
 
 vec3 getNormalForEnv(mat3 rotEnvMatrix, vec3 normal_inWorld, float materialSID) {
   vec3 normal_forEnv = rotEnvMatrix * normal_inWorld;
-  if (get_inverseEnvironment(uint(materialSID), 0u)) {
+  if (get_inverseEnvironment(materialSID, 0u)) {
     normal_forEnv.x *= -1.0;
   }
   return normal_forEnv;
@@ -218,7 +218,7 @@ vec3 getReflection(mat3 rotEnvMatrix, vec3 viewDirection, vec3 normal_inWorld, f
 #else
   vec3 reflection = rotEnvMatrix * reflect(-viewDirection, normal_inWorld);
 #endif
-  if (get_inverseEnvironment(uint(materialSID), 0u)) {
+  if (get_inverseEnvironment(materialSID, 0u)) {
     reflection.x *= -1.0;
   }
   return reflection;
@@ -231,10 +231,10 @@ vec3 IBLContribution(float materialSID, vec3 normal_inWorld, float NdotV, vec3 v
   vec3 iridescenceFresnel_dielectric, vec3 iridescenceFresnel_metal, float iridescence, float anisotropy, vec3 anisotropyDirection,
   float specularWeight, vec3 dielectricF0, float metallic, float diffuseTransmission, vec3 diffuseTransmissionColor, float diffuseTransmissionThickness)
 {
-  vec4 iblParameter = get_iblParameter(uint(materialSID), 0u);
+  vec4 iblParameter = get_iblParameter(materialSID, 0u);
   float rot = iblParameter.w;
   mat3 rotEnvMatrix = mat3(cos(rot), 0.0, -sin(rot), 0.0, 1.0, 0.0, sin(rot), 0.0, cos(rot));
-  ivec2 hdriFormat = get_hdriFormat(uint(materialSID), 0u);
+  ivec2 hdriFormat = get_hdriFormat(materialSID, 0u);
 
   vec3 normal_forEnv = getNormalForEnv(rotEnvMatrix, normal_inWorld, materialSID);
   vec3 reflection = getReflection(rotEnvMatrix, viewDirection, normal_inWorld, materialSID, perceptualRoughness, anisotropy, anisotropyDirection);
@@ -252,8 +252,8 @@ vec3 IBLContribution(float materialSID, vec3 normal_inWorld, float NdotV, vec3 v
 #endif
 
 #ifdef RN_USE_TRANSMISSION
-  vec3 attenuationColor = get_attenuationColor(uint(materialSID), 0u);
-  float attenuationDistance = get_attenuationDistance(uint(materialSID), 0u);
+  vec3 attenuationColor = get_attenuationColor(materialSID, 0u);
+  float attenuationDistance = get_attenuationDistance(materialSID, 0u);
   vec3 specularTransmission = getIBLVolumeRefraction(baseColor, normal_inWorld, viewDirection, cameraSID, materialSID, thickness, perceptualRoughness, ior, attenuationColor, attenuationDistance);
   diffuse = mix(diffuse, specularTransmission, transmission);
 #endif
