+++ src/webgl/shaderity_shaders/common/pointSprite.glsl
@@ -1,11 +1,11 @@
 // #ifdef RN_IS_POINTSPRITE
 
 vec4 position_inWorld = worldMatrix * vec4(a_position, 1.0);
-vec3 viewPosition = get_viewPosition(uint(cameraSID));
+vec3 viewPosition = get_viewPosition(cameraSID);
 float distanceFromCamera = length(position_inWorld.xyz - viewPosition);
-vec3 pointDistanceAttenuation = get_pointDistanceAttenuation(uint(materialSID), 0u);
+vec3 pointDistanceAttenuation = get_pointDistanceAttenuation(materialSID, 0u);
 float distanceAttenuationFactor = sqrt(1.0/(pointDistanceAttenuation.x + pointDistanceAttenuation.y * distanceFromCamera + pointDistanceAttenuation.z * distanceFromCamera * distanceFromCamera));
-float maxPointSize = get_pointSize(uint(materialSID), 0u);
+float maxPointSize = get_pointSize(materialSID, 0u);
 gl_PointSize = clamp(distanceAttenuationFactor * maxPointSize, 0.0, maxPointSize);
 
 // #endif
